Add kmesh attributes for PBC SCF class
